# cmd/api/docs/swagger.yaml
basePath: /v1
definitions:
  cerrors.IError:
    type: object
...
  elastic.LightContract:
    properties:
      address:
        type: string
      deploy_time:
        type: string
      network:
        type: string
    type: object
  elastic.ProjectStats:
    properties:
      contracts_count:
        type: integer
      first_deploy:
        type: string
      language:
        type: string
      last:
        $ref: '#/definitions/elastic.LightContract'
        type: object
      last_action:
        type: string
      last_deploy:
        type: string
      name:
        type: string
      tx_count:
        type: integer
      versions_count:
        type: integer
    type: object
  elastic.SameContractsResponse:
    properties:
      contracts:
        items:
          $ref: '#/definitions/models.Contract'
        type: array
      count:
        type: integer
    type: object
  elastic.SimilarContract:
    properties:
      added:
        type: integer
      address:
        type: string
      alias:
        type: string
      annotations:
        items:
          type: string
        type: array
      balance:
        type: integer
      consumed_gas_diff:
        type: number
      count:
        type: integer
      delegate:
        type: string
      delegate_alias:
        type: string
      diff:
        type: string
      entrypoints:
        items:
          type: string
        type: array
      fail_strings:
        items:
          type: string
        type: array
      fingerprint:
        $ref: '#/definitions/models.Fingerprint'
        type: object
      found_by:
        type: string
      hardcoded:
        items:
          type: string
        type: array
      hash:
        type: string
      id:
        type: string
      language:
        type: string
      last_action:
        $ref: '#/definitions/models.BCDTime'
        type: object
      level:
        type: integer
      manager:
        type: string
      migrations_count:
        type: integer
      network:
        type: string
      project_id:
        type: string
      removed:
        type: integer
      tags:
        items:
          type: string
        type: array
      timestamp:
        type: string
      total_withdrawn:
        type: integer
      tx_count:
        type: integer
    type: object
...
  handlers.Contract:
    properties:
      address:
        type: string
      alias:
        type: string
      annotations:
        items:
          type: string
        type: array
      balance:
        type: integer
      delegate:
        type: string
      delegate_alias:
        type: string
      entrypoints:
        items:
          type: string
        type: array
      fail_strings:
        items:
          type: string
        type: array
      fingerprint:
        $ref: '#/definitions/models.Fingerprint'
        type: object
      found_by:
        type: string
      hardcoded:
        items:
          type: string
        type: array
      hash:
        type: string
      id:
        type: string
      language:
        type: string
      last_action:
        $ref: '#/definitions/models.BCDTime'
        type: object
      level:
        type: integer
      manager:
        type: string
      migrations_count:
        type: integer
      network:
        type: string
      profile:
        $ref: '#/definitions/handlers.ProfileInfo'
        type: object
      project_id:
        type: string
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      timestamp:
        type: string
      total_withdrawn:
        type: integer
      tx_count:
        type: integer
    type: object
...
  handlers.Error:
    properties:
      message:
        example: text
        type: string
    type: object
  handlers.GetErrorLocationResponse:
    properties:
      end_col:
        type: integer
      failed_row:
        type: integer
      first_row:
        type: integer
      start_col:
        type: integer
      text:
        type: string
    type: object
...
  handlers.Operation:
    properties:
      amount:
        type: integer
      balance:
        type: integer
      balance_updates:
        items:
          $ref: '#/definitions/models.BalanceUpdate'
        type: array
      burned:
        type: integer
      counter:
        type: integer
      delegate:
        type: string
      destination:
        type: string
      destination_alias:
        type: string
      entrypoint:
        type: string
      errors:
        items:
          $ref: '#/definitions/cerrors.IError'
        type: array
      fee:
        type: integer
      gas_limit:
        type: integer
      hash:
        type: string
      id:
        type: string
      internal:
        type: boolean
      kind:
        type: string
      level:
        type: integer
      manager_pubkey:
        type: string
      mempool:
        type: boolean
      network:
        type: string
      parameters:
        type: object
      protocol:
        type: string
      public_key:
        type: string
      result:
        $ref: '#/definitions/models.OperationResult'
        type: object
      source:
        type: string
      source_alias:
        type: string
      status:
        type: string
      storage_diff:
        type: object
      storage_limit:
        type: integer
      timestamp:
        type: string
    type: object
  handlers.OperationResponse:
    properties:
      last_id:
        example: "1588640276994159"
        type: string
      operations:
        items:
          type: integer
        type: array
    type: object
...
  handlers.ProfileInfo:
    properties:
      subscribed:
        type: boolean
    type: object
...
  models.BCDTime:
    type: object
  models.BalanceUpdate:
    properties:
      category:
        type: string
      change:
        type: integer
      contract:
        type: string
      cycle:
        type: integer
      delegate:
        type: string
      kind:
        type: string
    type: object
...
  models.Contract:
    properties:
      address:
        type: string
      alias:
        type: string
      annotations:
        items:
          type: string
        type: array
      balance:
        type: integer
      delegate:
        type: string
      delegate_alias:
        type: string
      entrypoints:
        items:
          type: string
        type: array
      fail_strings:
        items:
          type: string
        type: array
      fingerprint:
        $ref: '#/definitions/models.Fingerprint'
        type: object
      found_by:
        type: string
      hardcoded:
        items:
          type: string
        type: array
      hash:
        type: string
      id:
        type: string
      language:
        type: string
      last_action:
        $ref: '#/definitions/models.BCDTime'
        type: object
      level:
        type: integer
      manager:
        type: string
      migrations_count:
        type: integer
      network:
        type: string
      project_id:
        type: string
      tags:
        items:
          type: string
        type: array
      timestamp:
        type: string
      total_withdrawn:
        type: integer
      tx_count:
        type: integer
    type: object
  models.Fingerprint:
    properties:
      code:
        type: string
      parameter:
        type: string
      storage:
        type: string
    type: object
  models.OperationResult:
    properties:
      allocated_destination_contract:
        type: boolean
      balance_updates:
        items:
          $ref: '#/definitions/models.BalanceUpdate'
        type: array
      consumed_gas:
        type: integer
      paid_storage_size_diff:
        type: integer
      storage_size:
        type: integer
    type: object
...
host: api.better-call.dev
info:
  contact: {}
  description: This is API description for Better Call Dev service.
  license: {}
  title: Better Call Dev API
  version: "1.0"
paths:
  /contract/{network}/{address}:
    get:
      consumes:
      - application/json
      description: Get full contract info
      operationId: get-contract
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Contract'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract info
      tags:
      - contract
...
  /contract/{network}/{address}/operations:
    get:
      consumes:
      - application/json
      description: Get contract operations
      operationId: get-contract-operations
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      - description: Last operation ID
        in: query
        name: last_id
        type: string
      - description: Timestamp
        in: query
        name: from
        type: integer
      - description: Timestamp
        in: query
        name: to
        type: integer
      - description: Expected OPG count
        in: query
        name: size
        type: integer
      - description: Comma-separated operations statuses
        in: query
        name: status
        type: string
      - description: Comma-separated called entrypoints list
        in: query
        name: entrypoints
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.OperationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract operations
      tags:
      - contract
...
  /contract/{network}/{address}/same:
    get:
      consumes:
      - application/json
      description: Get same contracts
      operationId: get-contract-same
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Requested count
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/elastic.SameContractsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get same contracts
      tags:
      - contract
  /contract/{network}/{address}/similar:
    get:
      consumes:
      - application/json
      description: Get similar contracts
      operationId: get-contract-similar
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/elastic.SimilarContract'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get similar contracts
      tags:
      - contract
...
  /operation/{id}/error_location:
    get:
      consumes:
      - application/json
      operationId: get-operation-error-location
      parameters:
      - description: Internal BCD operation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetErrorLocationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get code line where operation failed
      tags:
      - operations
  /opg/{hash}:
    get:
      consumes:
      - application/json
      description: Get operation group by hash
      operationId: get-opg
      parameters:
      - description: Operation group hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Operation'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get operation group
      tags:
      - operations
...
  /projects:
    get:
      consumes:
      - application/json
      description: Get all projects stats
      operationId: get-projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/elastic.ProjectStats'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Show projects
      tags:
      - projects
...
swagger: "2.0"